Fix connected-components
Kosaraju's algorithm needs to call `visit!` on all nodes, but the
original code only calls it on the nodes with inbound links
(the variable name is also misnomer).  Consequently, it returns
wrong results:

```
(connected-components '((1 2))) ⇒ ((1 2))
   ;; should be ((1) (2)) or ((2) (1))
```

I added some more tests, and use `lset=` to compare the result,
since the order of the result list and the order of elements in each
result list shouldn't matter.